Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2060 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ RadioButtonCmd_.h
@@ -55,7 +55,7 @@ class RadioButtonCmd : public Cmd
     const Widget	_button;			// commandWidget
     const Widget	_label;				// labelWidget
 
-    static u_int	_nitems;
+    static size_t	_nitems;
 };
 
 }
